[rfc] [ocaml] Enable warning "27" unused variable.
IMHO this warning is important to be on, specially in the upper layers
that are supposed to handle environments in "let style".

Unfortunately CAMPLX does not understand `[@@@ocaml.warning "-27"]` so
this PR is not ready for merge as `ml4` files will trigger the warning
in abundance due to unused `loc` generated by parsing rules.

Suggestions on how to disable the warning for `ml4`-generated files
are welcome.