Use old option --comp to avoid spurious failures.
It is very common these days that Travis will fail to fetch the latest
versions from the Ubuntu package repositories. In this case, it falls
back to using cached packages but the version of OPAM is older and does
not know about the --compiler option. So we are using the --comp option
instead for greater robustness.