[engine] Generalize type of evar maps, part I.
This is part 1 of an experiment to try to have evar_maps that contain
`EConstr`. To that purpose, we allow the evar maps to contain generic
terms.

As of now, we just do the minimal amount of changes as to preserve
compatibility, there still are some functions that should be fixed as
to be more general.