[travis] Remove some jobs that are run by CircleCI.
Travis has a lot of capacity issues these days (as well as spurious
failures), so we remove the jobs that are better handled by CircleCI
because of the dependency handling.

We may want to remove more jobs, since CircleCI handles the others
nicely, but let's try this as a first step.