Use travis_retry on apt-get update
Script modified from
https://unix.stackexchange.com/questions/175146/apt-get-update-exit-status

I stuck the code in "install" rather than "before_install" so that the
lint target didn't need to be changed.  I also haven't touched the
targets that add more packages; I'll leave that to someone who knows
more about the "&" and "*" syntax being used in the configuration.