Proofview: enter_one: add __LOC__ argument to get relevant error msg
The type discipline of the tactic monad does not distinguish between
mono-goal and multi-goal tactics. Unfortunately enter_one "asserts
false" if called on 0 or > 1 goals.  The __LOC__:string argument can
be used to make the error message more helpful (since the backtrace is
pointless inside the monad).

The intended usage is "Goal.enter_one ~__LOC__ (fun gl -> ..".

The __LOC__ variable is filled in by the OCaml compiler with the current
file name and line number.