Use whd-all on rigid-flex conversion.
This heuristic is justified by the fact that during a conversion check
between a flexible and a rigid term, the flexible one is eventually going
to be fully weak-head normalized. So in this case instead of performing
many small reduction steps on the flexible term, we perform full weak-head
reduction, including delta.

It is slightly more efficient in actual developments, and it fixes a corner
case encountered by Jason Gross.

Fixes #6667: Kernel conversion is much, much slower than `Eval lazy`.