Fix #6677: Critical bug with VM and universes
This bug was present since the first patch adding universe polymorphism
handling in the VM (Coq 8.5). Note that unsoundness can probably be
observed even without universe polymorphism.

This is an adaptation of a cherry-pick of d1288ee from v8.5 which
itself is an adaptation of c9f3a6c in master (PR #6713).