coq_makefile: Support "" as the prefix in _CoqProject
This fixes #6350 (and even comes with a test case).

Refering to other directories as `-R … ""` is maybe not best practice,
but some people out there do it, so as long as it does not cause too
much trouble, we can continue to support it.

(cherry picked from commit 00dfe2be53ddc84836e095f0a20a0efdebc87f50)