Remove the deprecation for some 8.2-8.5 compatibility aliases.
This was decided during the Fall WG (2017).

The aliases that are kept as deprecated are the ones where the difference
is only a prefix becoming a qualified module name.
The intention is to turn the warning for deprecated notations on.

We change the compat version to 8.6 to allow the removal of VOld and V8_5.