[build] Build checker generated files using a make rule.
Currently, `configure.ml` does copy/link some files from `kernel` to
`checker` in an ad-hoc way. Instead, it is preferable to add a copy
rule to make and let it handle the dependencies properly.

This also fixes a dependency bug in Windows, as files wouldn't be
properly refreshed if `configure` was not run each time.