Emergency fix for OSX packaging job on Travis.
Updates to homebrew packages can currently make this packaging job fail
at any stage of the release cycle. Also, we are leaving dangerously
because we depend on python2 and python3 at the same time, which is not
supported by homebrew. To make this more reliable, we should switch to a
Nix-based build infrastructure at least for macOS.