Refine a bit the decentralized merging process.
We make GitHub assign only principal maintainers as reviewers. This
reduces the level of noise (PRs with 10 code owners), and makes it easy
for the assignee to check if all reviews have been completed (all
reviewers in the list have to approve the PR, which was not the case
before if two reviewers were assigned for the same component).

This change means that when a principal maintainer submits a patch
touching the component they own, they should ask a review from the
secondary maintainer.