Congruence: Fixing a bug with native projections.
There is a code to turn constants denoting projections into proper
primitive projections, but it did not drop parameters.

The code seems anyway redundant with an "expand_projections" which is
already present in Cctac.decompose_term. After removal of this code,
the two calls to congruence added to cc.v work.