gitlab: fix environment for build template
When `build` was made to build the doc it dropped `-coqide opt` and
dropped the environment variables for building coqide.

The combination means that when the cache had lablgtk in
opam (installed by some other job) configure would pick it up but the
system package wouldn't be there causing a failure. When lablgtk isn't
in the cache everything was fine.

(cherry picked from commit 7cec7d312831f509a2d63e89173395ee428f46fa)