[ssr] fix delayed clears (fix #7045)
This commit executes `[IPatClear x; IPatView y]` as
`[IPaView y; IPatClear x] ` and fixes the computation of names used
in an intro pattern.


The complication is from the fact that, at the same time, we:
1 delay clears (to the end of the intro pattern).
  Eg `=> _ ->` on `forall x, x = 3 -> G`
2 let one reuse cleared names for introduction (everywhere after the clear)
  Eg `move=> {H} H`
3 let one use a cleared variable inside a term
  Eg `move=> {H} /H`
and the implementation uses `rename` behind the scenes

`move=> {x} pat` is semantically equivalent to
- `rename x _x_; move=> pat {_x_}` where `_x_` is a fresh
  inaccessible name if `x` is used in `pat`
- `move=> pat {x}` otherwise

What does *used* mean?

Used means "used as a name for introduction anywhere in the pattern".
Prior this commit, it meant "used as a name for introduction or cleared
in the fragment of the intro pattern following the current position".
It was wrong in two ways:
- Something "cleared" is not used
- in patterns like `[ {x} y | ] x` the name `x` was not considered as
  used because the code was looking only at `y`

Do we really need the generality of point 3? No.

Its implementation was broken/incomplete but in MathComp it was
working "by chance".

The idiom to support is simpler than 3:
`move=> .. {H}/H` where the clear is the step *just before* the view
it is like the `{}` clear "attribute" for rewrite rules.