Adapt the standard library and tests to use primitive corecursive records.
Most of the changes are actually due to the Coq "design pattern" that
CoInductive types do not generate induction schemes.

The only part of the stdlib that was really using them is the Stream library,
which is probably not much used in the wild.

Depending on the case, tests were either adapted, or removed if they were made
moot by the transition, e.g. if they relied on properties not holding with
negative coinductive types.