Distinguishing "sort" variables from "Type" variables in pretyping (bug #4328).
When an evar has to be instantiated by an unkown sort, a level >= Set
was generated, possibly prevening further unification of the sort with
Prop to happen.

This commit relaxes the constraint that such unknown sort is
necessarily >= Set.

This is done by allowing to declare variables >= Prop in the kernel. I
thus make here the assumption that the kernel graph algorithm, which
includes Prop as a node, is indeed designed to deal correctly with
Prop as a level.

The commit also includes a fix of define_evar_as_sort which was
declaring a sort variable with a "rigid" status, thus preventing it to
be unified.