Delay the computation of frozen evars in legacy unification.
This source of slowness has been observed in VST, but it is probably
pervasive. Most of the unification problems are not mentioning evars,
it is thus useless to compute the set of frozen evars upfront.

We also seize the opportunity to reverse the flag, because it is always
used negatively.

(cherry picked from commit 122a5aca4f7b21f65afece2c59e8529183713d71)