Micromega clean-up
We add .mli files, removed dead code and use standard combinators
instead of redefined ad-hoc ones in a few places.

A lot of cleaning still has to be done on this code: documenting the
interfaces, resolving the many abstraction leaks. I suspect there is
still a lot of code duplication.