[api] Merge `Declareops` into `Declarations`
This is an experiment in trying to get a more uniform API by merging
together definitions with their operating interfaces. The main
motivation to help the API documentation work (#7155), but this also
has a net effect on the number of `open` needed in the codebase and no
downside I can see.

Also see discussions in #6512, https://github.com/coq/ceps/pull/19 and
likely elsewhere.

IMO reducing the number of modules is a key point in order to make the
API accessible, if this PR finds some success the pattern:

```ocaml
open Foo
open Fooops
```

should hopefully go away.

cc: #7422