[gitlab] Add bleeding-edge flambda build.
We also introduce a bit more systematic job naming: `base/edge`.

In order to make the flambda switch selectable we update the Docker
image so all the dependencies are installed in that one.

Note the extra quote rule for the flambda parameters, but unless we
can assign arrays to Gitlab variables there is not a good way to do
this I'm afraid.

With this patch we are getting close to being able to remove most
builds from Travis.