[ci] Don't build lite versions of CI developments.
In the original Travis CI setup, the per-job time limit was an
issue. However, Gitlab has much improved this problem due to

 a) Coq not being built for each contrib,
 b) user-configurable time limit.

We thus disable the expensive builds from Travis:

 `fiat-crypto`, `formal-topology`, `geocoq`, `iris-lambda-rust`,
 `math-comp`, `unimath`, `vst`

and instruct Gitlab to build [`geocoq`, `math-comp`, `unimath`, `vst`]
in full.

We also update the `math-comp` script as the `odd-order` theorem lives
in a separate repository and it is a key CI case.